Move known related resources in known module
